chore(blog): simplify for maintenance
