+++ blog/.gitignore
@@ -1,5 +1,3 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
 # dependencies
 /node_modules
 /.pnp
@@ -8,12 +6,8 @@
 # testing
 /coverage
 
-# next.js
-/.next/
-/out/
-
-# production
-/build
+# build output
+/dist
 
 # misc
 .DS_Store
@@ -30,10 +24,8 @@ yarn-error.log*
 .env.test.local
 .env.production.local
 
-# vercel
-.vercel
-
-# generated rss feed
+# generated feeds and sitemap
 public/rss.json
 public/rss.xml
 public/atom.xml
+public/sitemap.xml
