--- blog/@types/post.d.ts
@@ -1,21 +1,0 @@
-export interface RawPost {
-  default: string;
-}
-
-export interface Post {
-  frontMatter: FrontMatter;
-  markdownBody: string;
-  slug: string;
-}
-
-export interface PostGroup {
-  [year: string]: Post[];
-}
-
-export interface PostProps {
-  siteTitle: string;
-  frontMatter: FrontMatter;
-  markdownBody: string;
-  previousPost?: Post;
-  nextPost?: Post;
-}
