--- blog/@types/site.d.ts
@@ -1,5 +1,0 @@
-export interface FrontMatter {
-  title: string;
-  author: string;
-  date: string;
-}
