--- blog/components/Layout.tsx
@@ -1,61 +1,0 @@
-import { ReactChildren, ReactChild, useEffect, useState } from 'react';
-
-import ProfileCard from '@components/ProfileCard';
-import SidebarMenu from '@components/SidebarMenu';
-import { Orb } from '@components/Orb';
-
-interface Props {
-  pageTitle: string;
-  showShare: boolean;
-  children: ReactChild | ReactChildren;
-}
-
-export default function Layout({ children, showShare }: Props) {
-  const [isMounted, setIsMounted] = useState(false);
-  const [menuOpenState, setMenuOpenState] = useState(false);
-
-  useEffect(() => {
-    setIsMounted(true);
-  }, []);
-
-  return (
-    <>
-      {isMounted && (
-        <>
-          <div className="container w-full mx-auto md:pt-10 md:max-w-6xl">
-            <span className="grid grid-cols-12">
-              <div className="visible md:hidden">
-                <SidebarMenu
-                  menuOpenState={menuOpenState}
-                  setMenuOpenState={setMenuOpenState}
-                />
-              </div>
-
-              <div className="hidden md:block">
-                <div className="col-start-1 col-end-4">
-                  <ProfileCard showShare={showShare} />
-                </div>
-              </div>
-
-              <div className="col-start-2 col-end-12 md:col-start-5">
-                <div className="visible md:hidden">
-                  <button
-                    className="pt-2.5"
-                    onClick={() => setMenuOpenState(!menuOpenState)}
-                    aria-label="toggle menu on smaller devices"
-                  >
-                    <div className="p-1">
-                      <Orb className="w-20 h-20" />
-                    </div>
-                  </button>
-                </div>
-
-                <section>{children}</section>
-              </div>
-            </span>
-          </div>
-        </>
-      )}
-    </>
-  );
-}
