+++ blog/index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Permanent+Marker&display=swap"
+      rel="stylesheet"
+    />
+    <link rel="icon" type="image/svg+xml" href="/icon.svg?v=3" />
+    <link rel="icon" type="image/x-icon" href="/favicon.ico?v=3" />
+    <link rel="apple-touch-icon" href="/icon.svg?v=3" />
+    <link rel="shortcut icon" href="/icon.svg?v=3" />
+    <meta name="theme-color" content="#fb923c" />
+    <title>Pete Hampton - Programmer</title>
+  </head>
+  <body class="font-sans leading-normal tracking-normal bg-gray-100 dark:bg-stone-900 dark:text-slate-200">
+    <div id="app"></div>
+    <script type="module" src="/src/main.tsx"></script>
+  </body>
+</html>
