+++ blog/package.json
@@ -2,52 +2,31 @@
   "name": "pjhampton-blog",
   "version": "1",
   "private": true,
+  "type": "module",
   "scripts": {
-    "build": "next build --webpack",
-    "clean": "rm -rf ./.next ./out",
-    "dev": "next dev --webpack",
-    "lint": "next lint",
-    "local": "next dev --webpack",
-    "scan": "npx react-scan@latest http://localhost:3000",
-    "start": "next start",
+    "build": "node scripts/generate-rss.js && node scripts/generate-sitemap.js && vite build && node scripts/generate-static-pages.js",
+    "clean": "rm -rf ./dist",
+    "dev": "vite",
+    "preview": "vite preview",
     "prettify": "npx prettier . --write"
   },
   "dependencies": {
-    "@next/bundle-analyzer": "^14.0.1",
-    "feed": "^4.2.2",
-    "framer-motion": "^11.5.6",
-    "gray-matter": "^4.0.3",
-    "next": "^16.1.6",
-    "next-seo": "^6.6.0",
-    "next-themes": "^0.2.1",
-    "postcss-preset-env": "^10.0.3",
-    "raw-loader": "^4.0.2",
-    "react": "^18.3.1",
-    "react-burger-menu": "^3.0.9",
-    "react-dom": "^18.3.1",
-    "react-fade-in": "^2.0.1",
-    "react-icons": "^5.3.0",
+    "preact": "^10.25.4",
+    "preact-router": "^4.1.2",
     "react-markdown": "^8.0.7",
-    "react-syntax-highlighter": "^15.5.0",
-    "rss": "^1.2.2",
-    "sass": "^1.79.2",
-    "sharp": "^0.33.5"
+    "react-syntax-highlighter": "^15.5.0"
   },
   "devDependencies": {
+    "@preact/preset-vite": "^2.9.3",
     "@tailwindcss/postcss": "^4.1.18",
     "@types/node": "^22.5.5",
-    "@types/react": "^17.0.35",
-    "@types/react-burger-menu": "^2.8.7",
     "@types/react-syntax-highlighter": "^15.5.9",
-    "@types/rss": "^0.0.32",
-    "@types/webpack-env": "^1.18.5",
-    "autoprefixer": "^10.4.20",
-    "browserslist": ">=4.23.3",
-    "eslint": "^8",
-    "eslint-config-next": "^16.1.1",
+    "feed": "^5.2.0",
+    "gray-matter": "^4.0.3",
     "postcss": "^8.4.47",
     "prettier": "3.3.3",
     "tailwindcss": "^4.1.18",
-    "typescript": "^5.6.2"
+    "typescript": "^5.6.2",
+    "vite": "^6.0.0"
   }
 }
