+++ blog/postcss.config.js
@@ -1,8 +1,5 @@
-module.exports = {
+export default {
   plugins: {
-    '@tailwindcss/postcss': {
-      config: './tailwind.config.js'
-    },
-    autoprefixer: {}
+    '@tailwindcss/postcss': {}
   }
 };
