+++ blog/src/components/Layout.tsx
@@ -1,0 +1,62 @@
+import { useState, useEffect } from 'preact/hooks';
+import type { ComponentChildren } from 'preact';
+
+import ProfileCard from './ProfileCard';
+import SidebarMenu from './SidebarMenu';
+import { Orb } from './Orb';
+
+interface Props {
+  pageTitle: string;
+  showShare: boolean;
+  children: ComponentChildren;
+}
+
+export default function Layout({ children, showShare }: Props) {
+  const [isMounted, setIsMounted] = useState(false);
+  const [menuOpenState, setMenuOpenState] = useState(false);
+
+  useEffect(() => {
+    setIsMounted(true);
+  }, []);
+
+  return (
+    <>
+      {isMounted && (
+        <>
+          <div className="visible md:hidden">
+            <SidebarMenu
+              menuOpenState={menuOpenState}
+              setMenuOpenState={setMenuOpenState}
+            />
+          </div>
+
+          <div className="container w-full mx-auto md:pt-10 md:max-w-6xl">
+            <span className="grid grid-cols-12">
+              <div className="hidden md:block">
+                <div className="col-start-1 col-end-4">
+                  <ProfileCard showShare={showShare} />
+                </div>
+              </div>
+
+              <div className="col-start-2 col-end-12 md:col-start-5">
+                <div className="visible md:hidden">
+                  <button
+                    className="pt-2.5"
+                    onClick={() => setMenuOpenState(!menuOpenState)}
+                    aria-label="toggle menu on smaller devices"
+                  >
+                    <div className="p-1">
+                      <Orb className="w-20 h-20" />
+                    </div>
+                  </button>
+                </div>
+
+                <section>{children}</section>
+              </div>
+            </span>
+          </div>
+        </>
+      )}
+    </>
+  );
+}
