+++ blog/src/styles/sidebar.css
@@ -1,0 +1,60 @@
+/* Custom sidebar menu - replaces react-burger-menu */
+
+.sidebar-menu {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 300px;
+  height: 100%;
+  background: white;
+  text-align: center;
+  padding: 2.5em 1.5em 0;
+  font-size: 1.15em;
+  z-index: 1100;
+  transform: translateX(-100%);
+  transition: transform 0.3s ease;
+  overflow-y: auto;
+
+  .dark & {
+    background: rgb(24 24 27);
+  }
+}
+
+.sidebar-menu--open {
+  transform: translateX(0);
+}
+
+.sidebar-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.3);
+  z-index: 1000;
+  opacity: 0;
+  visibility: hidden;
+  transition: opacity 0.3s ease, visibility 0.3s ease;
+}
+
+.sidebar-overlay--open {
+  opacity: 1;
+  visibility: visible;
+}
+
+.sidebar-close {
+  position: absolute;
+  top: 12px;
+  right: 12px;
+  background: none;
+  border: none;
+  color: #bdc3c7;
+  font-size: 28px;
+  cursor: pointer;
+  padding: 4px 8px;
+  line-height: 1;
+}
+
+.sidebar-close:hover {
+  color: #a9beef;
+}
