+++ blog/src/utils/theme.tsx
@@ -1,0 +1,50 @@
+import { createContext } from 'preact';
+import { useContext, useState, useEffect, useCallback } from 'preact/hooks';
+import type { ComponentChildren } from 'preact';
+
+interface ThemeContextType {
+  resolvedTheme: string;
+  setTheme: (theme: string) => void;
+}
+
+const ThemeContext = createContext<ThemeContextType>({
+  resolvedTheme: 'light',
+  setTheme: () => {}
+});
+
+export function ThemeProvider({ children }: { children: ComponentChildren }) {
+  const [resolvedTheme, setResolvedTheme] = useState(() => {
+    if (typeof window !== 'undefined') {
+      const stored = localStorage.getItem('theme');
+      if (stored) return stored;
+      return window.matchMedia('(prefers-color-scheme: dark)').matches
+        ? 'dark'
+        : 'light';
+    }
+    return 'light';
+  });
+
+  const setTheme = useCallback((theme: string) => {
+    setResolvedTheme(theme);
+    localStorage.setItem('theme', theme);
+  }, []);
+
+  useEffect(() => {
+    const root = document.documentElement;
+    if (resolvedTheme === 'dark') {
+      root.classList.add('dark');
+    } else {
+      root.classList.remove('dark');
+    }
+  }, [resolvedTheme]);
+
+  return (
+    <ThemeContext.Provider value={{ resolvedTheme, setTheme }}>
+      {children}
+    </ThemeContext.Provider>
+  );
+}
+
+export function useTheme() {
+  return useContext(ThemeContext);
+}
